[FLINK-5619] Add numStateEntries() method for all keyed backends
This also adds a test for this in StateBackendTestBase
